+++ packages/hexboosting-sdk/src/constants.ts
@@ -1,0 +1,5 @@
+import { PublicKey } from "@solana/web3.js";
+
+export const PROGRAM_ID = new PublicKey(
+  "hexbnKYoA2GercNNhHUCCfrTRWrHjT6ujKPXTa5NPqJ"
+);