+++ packages/account-postgres-sink-service/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.17](https://github.com/helium/helium-program-libary/compare/v0.6.16...v0.6.17) (2024-01-19)
+
+**Note:** Version bump only for package @helium/account-postgres-sink-service
+
+
+
+
+
+## [0.6.16](https://github.com/helium/helium-program-libary/compare/v0.6.15...v0.6.16) (2024-01-16)
+
+
+### Features
+
+* **#543:** Add transaction webhook capability to account-postgres-sink ([#544](https://github.com/helium/helium-program-libary/issues/544)) ([a0d7bd9](https://github.com/helium/helium-program-libary/commit/a0d7bd9cbf6e5a28c467d8df67c9702cef705dae)), closes [#543](https://github.com/helium/helium-program-libary/issues/543)
+
+
+
+
+
+## [0.6.15](https://github.com/helium/helium-program-libary/compare/v0.6.14...v0.6.15) (2024-01-12)
+
+**Note:** Version bump only for package @helium/account-postgres-sink-service
+
+
+
+
+
+## [0.6.14](https://github.com/helium/helium-program-libary/compare/v0.6.13...v0.6.14) (2024-01-11)
+
+**Note:** Version bump only for package @helium/account-postgres-sink-service
+
+
+
+
+
+## [0.6.13](https://github.com/helium/helium-program-libary/compare/v0.6.12...v0.6.13) (2024-01-05)
+
+**Note:** Version bump only for package @helium/account-postgres-sink-service
+
+
+
+
+
+## [0.6.12](https://github.com/helium/helium-program-libary/compare/v0.6.11...v0.6.12) (2024-01-02)
+
+**Note:** Version bump only for package @helium/account-postgres-sink-service
+
+
+
+
+
+## [0.6.11](https://github.com/helium/helium-program-libary/compare/v0.6.8...v0.6.11) (2023-12-27)
+
+**Note:** Version bump only for package @helium/account-postgres-sink-service
+
+
+
+
+
 ## [0.6.10](https://github.com/helium/helium-program-libary/compare/v0.6.5...v0.6.10) (2023-12-22)
 
 **Note:** Version bump only for package @helium/account-postgres-sink-service